split tests to mitigate bottlenecks
